Implemented a chunk-based parser.
The parser will now accept a series of string chunks.  Parsing will end
when a nil is received.

ftcsv.parse now accepts a simple iterator function as a first parameter.
The function should return the next string chunk in the series when it
is called, returning nil when the input ends.